WIP add AVRs to ArchiveBlock proto
This is an exploratory branch based on the idea that, we should
store the AVRs that support the block signatures in Amazon S3 storage,
in addition to storing them with the watchers.

This PR adds the verification report to the ArchiveBlock object.

That change requires moving the `BlockData` object out of `mc-transaction-core`
because it isn't acceptable for `mc-transaction-core` to depend on
`mc-attest-core`, that could make it very difficult to create hardware
wallets and such without cross-compiling `mc-attest-core` and all
its dependencies.

It also isn't good for `mc-attest-core` to depend on `mc-transaction-core`.

We decided that we should think of `BlockData` as a ledger type, since
it's really an S3 protobuf type corresponding to the `ArchiveBlock`
in `external.proto`. We moved it to `mc-ledger-types` crate, and renamed
it to `ArchiveBlock` to match the protobuf.

We then made
- `mc-ledger-db` have storage for verification reports, and de-duplicate
  these reports.
- `mc-ledger-sync` propagates any found verification reports to callers
  etc.
- More such plumbing like this.

It's not clear that we needed to extend `mc-ledger-db` in this way
to meet the goal, but it seems like it doesn't hurt and we also wanted
to see how much work this is.

---

Feedback on the direction in this PR is greatly appreciated. What
we learned from developing this will also be helpful for trying to
implement Guaranteed Timestamps.

(Guaranteed Timestamps means roughly, the Ledger also tracks the
block signatures not just of its own node, but also of peer nodes,
when blocks are obtained from the catchup process. Those peer
signatures are also added to `ArchiveBlock`. They are stored
separately from the node's own signature to avoid ambiguity. The
benefit of this is that there will not be a scenario going forward
that you get blocks with no signature and no timestamp if you download
from S3 -- currently fog has to rely on the watcher for timestamps,
which is more complicated.)