Fix ingest not publishing metrics #1081
It is reported that when ingest dies (perhaps Azure restarts the VM)
and then is restarted and begins in the idle state, it doesn't publish
any metrics saying that it is in the idle state.

Initially, I thought that we must be missing a line in the worker to
tell it to publish metrics or something, but as I dug into that,
it seems that that is happening in the grpc admin API and the server
does not need to regularly push out metrics actively on its own.

I dug into what is setting the metrics, and it looks that when
the server changes mode, we update the metrics. However, when the
controller state is first initialized, that does not trigger any
metrics to be set. (I.e. we are trying to maintain an invariant
that the controller state value is always published to the metrics,
but that invariant is not established in the constructor.)

I think this change should fix the issue.

I also fixed up some documentation that got mangled during formatting