Make a "canary" mode of the fog-test-client target
This refactors the `fog-test-client` and makes a new mode for it,
which is appropriate for running as a "canary" in prod.
The canary can both alert when transactions are failing, and
publish tx timing stats for prod.

The new options are:
`--admin-listen-uri` specifies a URI at which to host the AdminAPI.
This endpoint exposes the prometheus metrics, so that they can be
tracked in grafana (and then alerted on if desired.)
When admin-listen-uri is specified, the process runs tests "continously"
pausing for `transfer_period` in between, which defaults to 15 minutes.

`--transfer-period` can specify an alternate frequency for sending
test transfers.

The prometheus metrics include:
* A histogram for how long until a transaction is received by
  the receiving client.
* A histogram for how long until a transaction is confirmed as
  having landed in the blockchain, by the sending client.
* A count for how many times either of these times exceeds the configured
  "deadline".
* A count for how many transfers are successful
* A count for each possible failure mode of a transfer, including,
  either of the clients returning any kind of error, or returning a bad
  balance, or the transaction being rejected, or expiring, or the memos
  being incorrect or invalid.

The internal refactor includes:
* `TestClient` is restructured to reuse code across the "fixed number
  of transfers" and the "continuous transfers" modes of operation.
* `TestClient` is restructured to use threads, so that we can have one
  thread measuring the tx receipt time, while a different thread is measuring
  the tx confirmed time, otherwise we can't measure both independently.
* `TestClient` always updates the prometheus counters, even if they aren't
  being published, because it's simple and doesn't break anything.
* Many `TestClient` panics are replaced with logging and returning errors,
  so that the canary mode of operation will not crash in the face of errors.
* `SamplePaykit` uses the nice URI types instead of strings

We thought about making the canary be a different binary from the test-client,
but it seems like it was better to reuse the code, even though it makes the
test client code a little more complicated than before. This way we will get
some test coverage on the canary by running the fog-test-client in CD.