update pre-commit hook to remove duplicates when formatting (#1093)
sometimes, when there is a really large diff, the pre commit hook
formatting step takes 20+ minutes. I believe that this is because
it formats the same file hundreds of times if there are a hundred
different segments of the diff related to that file. This minor
change removes duplicates before running cargo fmt.