Make the sample-keys program emit b58-formatted public address files as well
This is helpful if you want to fund an account created using the
`sample-keys` program. The main reason to do that is that the `sample-keys`
program is easy to run in automation rather than having spreadsheets or
tables of keys, and it produces the root-identity files which our targets like
bootstrap, fog-distribution and the test-client are compatible with.

However, the b58 format is what full-service and other wallets use,
and it's not super easy to convert from the b58 format to this format.

An alternative here is that we could make the `.pub` file be the b58
format and kill the old proto format, but that's a breaking change
and I'm not sure what parts of infrastructure are still using these
files, so I'd rather not do that and just remove the `.pub` file
entirely at some later date.