Add retries to many fog enclave connections using retry crate (#1163)
* Add retries to many fog enclave connections using retry crate

This addresses many old tickets, and I developed this while trying
to figure out why the test client failed with seemingly retriable errors

There is a large amount of tech debt to resolve the discrepancy
between mc-connection and mc-fog-enclave-connection. They are mostly
supposed to be doing the same thing, but `mc-fog-enclave-connection`
is generic over a protobuf client object that could be ledger or view
(or consensus user tx conceivably). OTOH the mc-connection objects
have been used in production and support more features like the credentials
provider, and before now, doing retries with the retry crate for
connection / attestation issues.

This does not attempt to resolve that issue for now, but it does port
forward the retry aspect of this from "sync_connection".

I decided not to do it in exactly the same way sync_connection, instead
I decided that `enclave_connection` should have a helper function that
converts its result into a `retry::OperationResult`, so that the caller
can use the `retry` call, and we don't need to pass the retry Iterator
into this call. To me this seems simpler and cleaner than what is done
in `mc-connection`, and does not involve any rust macros to get it to work.

This will hopefully make unit tests and CD more stable, as well as making
the test client more stable.

* fix clippy