Use slip10 accounts instead of legacy root entropy in internal testing
This migrates the `mc-util-keyfile` crate to write private key json
files containing slip10 root entropies instead of the legacy root entropy.
The legacy root entropies are deprecated but they must still exist
in some capacity because there are accounts with money on them IRL.

This will simplify the effort to get uniform integration tests
because many wallet products may never implement root entropies.

The bootstrap program does not need to be touched because it only
reads the public key files.

The keygen program is touched, fog-distro is touched, the sample
paykit is touched, the test-client is touched, slam is touched,
but most of the changes are very localized.

The main thing that would make this not work is if someone somewhere
is storing keyfiles in a persistent way, however this is unlikely
and whatever system that is likely needs to be reworked anyways.