fixup what happens when Tx hangs in limbo
When a Tx hangs in limbo (past deadline), we immediately report unhealthy
and log all available context for investigation.

This makes a TxInfo object which is shared as part of the test client
state and updated as the tx progresses, and which can be logged nicely
with whatever info we have.

This removes the log guard object which would have been more complicated
to integrate with this.