remove rpc_enclave_err, fix a bunch of error -> rpc conversions (#1252)
* remove rpc_enclave_err, fix a bunch of error -> rpc conversions

This commit starts from the idea that `rpc_enclave_err` is wrong.
Usually, an enclave error is a complex enum containing numerous
possible failure modes. Some of these may be SGX errors that are
not retriable and may represent internal errors. However,
attestation errors must be mapped to "permission denied" to trigger
re-attestation on the clients. Deserialization errors might represent
invalid arg, or might represent an error internal to the server
depending on which payload couldn't be deserialized.

The `rpc_enclave_err` thing was created during very early days
in the project when we were just trying to get things working
and developers didn't want to spend a lot of time trying to do
correct status codes for these errors. It is very convenient to
have a single function for reporting all "enclave errors" but
unfortunately this is usually a trap for the developer nowadays.

Therefore we attempt to remove `rpc_enclave_err` and then review
what code is broken by that and try to do it more nicely.

* consensus peer api service now matches on the enclave error and
  returns permission_denied or internal error
* ingest peer api service now matches on the enclave error and
  returns permission_denied or internal error

While we were at it, we improved a bunch of other error mappings
in fog ingest, for calls triggered by the fog ingest client, like
activate, new-keys, sync-keys-from-remote, etc.

This is not extremely important now, but may help when we tackle
automated failover for fog ingest and have automated systems calling
these APIs which need to know when to retry etc.

* fix error context strings

* Update fog/ingest/server/src/ingest_service.rs

Co-authored-by: Eran Rundstein <eran@rundste.in>

* eran match statement suggestions

* more simplifications

Co-authored-by: Eran Rundstein <eran@rundste.in>