Add support for k8s readiness checks in fog view and fog ledger (#1336)
* Add support for k8s readiness checks in fog view and fog ledger

In kubernetes, services have both a "liveliness" status and a
"readiness" status. Servers may be live if they are running, but
they are not ready serve requests until they have loaded a bunch
of data. This well represents the fog servers, since loading data
into ORAM takes time.

https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-startup-probes/#define-a-grpc-liveness-probe

We already have readiness checks in our k8s configs, but fog view
and fog ledger aren't exposing this information to k8s right now.
This commit makes them expose this by returning the "NOT_SERVING"
status until such point as they they think they have loaded all
(or nearly all) data into ORAM.

This modifies the way these `fog` servers are using the
`mc-util-grpc` health service. It creates a handle `ReadinessIndicator`
which can be passed around in servers so that they can signal when
they are ready.

Some parts of this are not perfect -- for `fog-ledger` it would be
better to couple the readiness check with a readiness check to `mobilecoind`,
so that we don't indicate that we are ready to receive traffic until
`mobilecoind` thinks it has synced the ledger at some point. It seems
like it is a substantial improvement that will get us most of the benefit
of implementing readiness checks though.

* make readiness indicator conversion more idiomatic

* fix tests build

* address review comments