In digestible crate, allow that `string` and `bytes` can be omitted
In digestible crate, part of the design is to allow that new `optional`
and `repeated` fields can be added to structs corresponding to protobufs,
without changing the hashes of old objects. By making the hash algorithm
similar to protobuf in this regard, we make it easier to evolve the
blockchain without breaking backwards compatibility.

Unfortunately, there are some gaps still in what kinds of schema
evolution digestible allows vs. what protobuf allows. Protobuf allows
that new bytes and string fields can be added to a struct without a
breaking change, but digestible doesn't, because it doesn't currently
omit empty bytes or string fields when omitting them would be allowed.
We didn't think about this when implementing the allow_omit stuff,
because we were focused on containers -- but bytes and strings are
primitives in protobuf.

Currently, NO primitives can be added unless they are Optional in rust.
Wrapping things in Optional<...> also works for Digestible, but it doesn't
always work well for things like prost.

This commit makes it so that strings and byte vectors have the "omit if empty"
behavior and can be added easily to our prosty structures.

I believe that there are currently no string objects in our blockchain.
I believe that there are no byte objects that are
being hashed by digestible that are not either fixed size, or wrapped in
optional.

This commit does not adjust the behavior of integers like u64 and u32,
which could also arguably be omitted if zero. However we have a fine
workaround for that in place which is the `digestible(omit_when = 0)`
which Eran made. Unfortunately it is fairly difficult to make that work
when the value to be omitted is not a literal.

If we decide that it is safe to merge this, this will simplify the changes
needed to support masked token ids.

There are other approaches we could take like
* adding more attributes to digestible, or,
* representing this type using `Option<GenericArray<u8, U4>>`.

But the latter requires a bunch of repr-bytes stuff with `prost` and `serde`
macros which in the long run we would like to reduce.

Dan Burkert was non-committal about supporting custom `prost::Message`
implementations: https://github.com/tokio-rs/prost/pull/421#issuecomment-758867131
and suggested that this is an "at your own risk" undertaking not covered
by semver guarantees.

Additionally, we considered it surprising that `Vec<u8>` behaves differently
from `Vec<T>` around whether it is omitted when empty. By merging this we would
eliminate that footgun.