Fixup transaction validation for 1.2.0 (#1460)
* Fixup transaction validation for 1.2.0

Per James, I actually got the release plan wrong for the RTH stuff.
The idea was supposed to be, update all readers first, in 1.2.0,
and only in the next release may writing memos be allowed.

This means that, transaction validation should enforce for now
that memos do not exist in the chain.

In MCIP #26 we may introduce a mechanism to allow turning on the
memos. That is not yet implemented. We will continue to work on that
but by merging this to master now, we make it possible to branch
for release now, and create clients capable of reading memos, while
we could later make a release that accepts writing memos in the
future.

We can discuss separately if the scope for 1.2.0 should be to
include a `block_version`-aware `TransactionBuilder`, it's possible
that it should be and then we aren't ready to branch. But then this
commit is still a fine basis for future iteration.

* Update transaction/core/src/validation/validate.rs

Co-authored-by: Remoun Metyas <remoun.metyas@gmail.com>

* cargo fmt

* fix tests that broke when we fixed other tests

* fixup tests (there is only one output in this tx)

Co-authored-by: Remoun Metyas <remoun.metyas@gmail.com>