add more run-time checks against block version in validation and building
block version is "guaranteed" by the class invariant not to exceed
the max, but strange things can happen if there is unsafe code or
a bug in serde or something. these checks will not be optimized out
in my understanding, because, the compiler has to assume that
class invariants can be violated by transmute -- transmute may be
unsafe but it does not make your program ill-formed.

reviewers seemed to agree that this is fine, and it can't hurt anything,
the worst case is it gets stripped out by optimizer.