Elevate watcher log level for sync errors
In the 2021-12-25 mainnet outage, no watcher logs were found
indicating trouble syncing attestation signatures. These logs do
indeed exist, but the log level was below that of what the log
shipper was able to pickup. Has been raised to "error" if launched
from mobilecoind with a fallback to debug if launched from the
standalone binary