Remove struct `ViewKey`, add a `view_key_match` function to `TxOut` (#1649)
* Remove struct `ViewKey`, add a `view_key_match` function to `TxOut`

Fixes #1594, #1600

`ViewKey` is removed because, while it is part of the original
cryptonote paper, it is not actually very useful. It is not needed
to actually determine if a `TxOut` is yours -- only the account
`view_private_key` is needed for that, and all the software works
that way. In another project which is attempting to support
"view-only accounts", the `ViewKey` struct is not even used. This
is partly because `ViewKey` is not needed, and partly because
`ViewKey` is a subaddress-specific value, and not account-wide,
the way the view private key is.

The `view_key_matches_output` function is removed from transaction
core, because it used `view_key`. No one was actually using this
function, again, because what it is actually doing is testing if
a TxOut matches to a specific subaddress. There were some tests
that use it, but they could easily be refactored to use other
functions.

`view_key_match` function was added to `TxOut`, which takes the
view private key, and attempts to match it against the TxOut.
This allows to simplify code in various places.

* fix clippy

* fix a clippy