add tests for untrusted tampering with data and metadata (#1732)
This is a test-driven approach to investigating and fixing the issue
https://github.com/mobilecoinfoundation/mobilecoin/issues/1576

These tests confirm that the trusted side indeed does not always
detect when untrusted tampers with the memory, at this revision
I have these results:

```
running 5 tests
test helpers::exercise_oram_storage_clear_data ... FAILED
test helpers::exercise_oram_storage_hammer_data ... FAILED
test helpers::exercise_oram_storage_hammer_metadata ... ok
test helpers::exercise_oram_storage_clear_metadata ... FAILED
test helpers::exercise_oram_storage_shims ... ok
```

So the shims work normally when the memory is not tampered with,
and the code detects when the untrusted metadata is hammered.
But it does not detect when it is cleared, or when the data is
tampered with.

This is likely because of the issue point out in #1576, in the
next commits we will fix it and get all these tests to pass.

WIP try to fix trusted merkle root check

Fix typo in readme

Fix tests by using a branch that shares a trusted merkle root.

Adding clarifying names to indices used in the test to better see what it is doing

Co-authored-by: Chris Beck <beck.ct@gmail.com>