Make GetLastBlockInfo return the configured block version, and use it
We realized that there is a chicken and egg problem for many clients:

* For mobilecoind and for fog clients, they were getting block version
  from their local copy of the ledger, or fog-ledger's copy of the ledger
* However, when we upgrade consensus and increase block version, the
  first transaction takes place when the ledger has the old block version
  but the consensus config is the new block version.
* There has to be at least one client capable of determining the configured
  block version and making a transaction according to those rules.

To fix this, James suggested that we use the `GetLastBlock` call, where
most clients are getting the minimum fees right now. We added a field
on this response that returns the configured block version, so this is
a good source of truth for the mentioned situation. Then we made mobilecoind,
fog-distro, and fog sample paykit all track the responded value and incorporate
it into their caches, so they are tracking the max of the ledger and this
response value.

This will also ease CD work because fog distro doesn't need to be told
the block version any more, and none of these other clients will have
problems after we deploy a new version.

We may want to backport this to 1.2.0 also.