Implement ledger sink/pass through component
This component is meant to sit downstream a component that has
verified a block and designed to trust the upstream blocks and
sink them into the db. After the sink the block stream components
pass through so other stream components can use them. Optionally
there is a method only sinking the blocks.