Add counter to track the gRPC status code in gRPC helper method
The method above is used to make a variety of gRPC calls. Operations
expressed the desire to be able to track different gRPC status codes
in FOG and consensus to better be able to put alerts on outages. As
many FOG & Consensus gRPC calls pass through this method, this will
track status codes for a wide variety of gRPC method calls across FOG
and consensus allowing ops to be able to establish alerts and better
profile performance.