change token id from u32 to u64
This is done because it is conceivable that we will have more than
4 billion tokens on our chain, particularly if this mechanism is
used to create NFTs or something someday.

There is very little reason from an engineering point of view to
try to save space by using a u32 instead of a u64 -- u32's do
sometimes run out in practice, and a u64 will not here.