move validation tests to be an integration test, remove adapt_hack
This works because, when cargo compiles in-line unit tests, they are
compiled into the module of the code under test, but when cargo
compiles integration tests (in tests directory), that is considered
a separate module (effectively a separate crate, just without its
own cargo toml).

This fixes the adapt hack issue because now we no longer need to
compile the `mc-transaction-core` crate in test mode, which also
references the `mc-transaction-core` crate which `mc-ledger-db`
depends on. The types in the test code always refer to the types
in the same version of `mc-transaction-core` that `mc-ledger-db`
is using.