Make the pre-commit hook check for copyright statement in rust files
This helps my workflow because I often don't want to be bothered with
this while I'm actually working, but at the point of commit is a
good place to check. This tool will fix it if the copyright are missing.

This will avoid the need for eran to put "nit: copyright" on every
PR that I make nowadays