fix formatting check in pre-commit
There is a subtle issue where, whenever formatting is necessary,
the `$diff` variable starts with a blnak line. This is because
we add things to `$diff` by setting `$diff = $diff\n$next`, so
the first item always sticks a new line in there even though it
is the first one.

(I discovered this first when doing the copyright check work.)

I think this means that when doing a formatting check, we do a
`cargo fmt` call over the entire repo first, then over each individual
file in the diff. It will probably be a lot faster if we don't do
the entire repo first, and only do the individual files of the diff,
which I think was the original intent.