Make mobilecoind expose network fees via an API call
This makes mobilecoind hit consensus' `get_last_block_info` as part
of the `GetNetworkStatus` call. It forwards a subset of that information
in response.

There is some argument that it should forward an entire `LastBlockInfo`
response object without modification -- I decided not to do that
because mobilecoind's proto doesn't currently depend on the consensus
api proto, only the `external.proto` and `blockchain.proto`. Feel
free to tell me to that dep is okay though.

This will allow mobilecoind python scripts and other clients of
mobilecoind to know the current (dynamic) network fees and block
version and such.