Make mobilecoind expose network fees via an API call (#2060)
* Make mobilecoind expose network fees via an API call

This makes mobilecoind hit consensus' `get_last_block_info` as part
of the `GetNetworkStatus` call. It forwards a subset of that information
in response.

There is some argument that it should forward an entire `LastBlockInfo`
response object without modification -- I decided not to do that
because mobilecoind's proto doesn't currently depend on the consensus
api proto, only the `external.proto` and `blockchain.proto`. Feel
free to tell me to that dep is okay though.

This will allow mobilecoind python scripts and other clients of
mobilecoind to know the current (dynamic) network fees and block
version and such.

* Per Eran, make polling_network_state track BlockInfo, use this in mobilecoind

This allows us to eliminate a bunch of calls to get block info, which
were being made to get the network_block_version and minimum fees.

* Make mobilecoind forward the whole LastBlockInfoResponse object

Per Eran

* missing docs, use `max_by_key` instead of `reduce`