slam as a service
This extends `mobilecoind-dev-faucet` with "slam" functionality.
By hitting an endpoint with curl, you can trigger it to spawn a
bunch of threads and build and submit tx's in rapid succession.
Then you can observe the performance of a deployed network when
it is under heavy load.

This is much like how `mc-fog-distribution` works, which is itself
a fork of an earlier tool "slam".

---

It may seem very odd to add load testing functionality to a program
that is currently a faucet. A faucet normally has the goal of giving
users a way to request small amounts of money for testing, and load
testing is not directly relevant to this.

The motivation here has more to do with problems and short comings
of the historical slam tool / nowadays fog-distro.

Slam / `mc-fog-distribution` submit transactions to the network,
but these tools are not "wallets". They simply take the origin block,
load keyfiles from the bootstrap procedure, and scan the origin
block asserting that certain account keys own certain TxOut's. Then,
many threads are spawned and transactions are built and submitted
which spend all these TxOut's as rapidly as possible.

This was created in early days when we wanted to stress-test the
network as simply as possible, to try to break consensus, and to
try to get max tx/s metrics.

Later, we needed a way to fund fog accounts for testing in continuous
deployment. It is not possible to bootstrap money to fog accounts
because bootstrap runs before the fog public key is known.
We were already running slam in continuous deployment, so the next
simplest thing would be to make slam able to fund fog accounts.
Fog distro was born, with a dual mandate, of both replacing slam
for load testing, and funding fog accounts.

Because fog distro was highly probabilistic (it fires-and-forgets
transactions rather than checking which ones land or don't and
resubmitting the failed ones), sometimes SDK tests would fail
because a test account had a zero balance. Over time various
people have jumped in to try to make it more reliable to fix
CD at various points.

---

In the latest release cycle, we discovered a lot of shortcomings
of this tool.

* fog-distro cannot be used to distribute the custom tokens, because
  those tokens cannot be bootstrapped, and fog-distro is intimately
  tied to the bootstrapped ledger. But we want to run tests with
  those tokens with fog clients.
* fog-distro cannot be run more than once for a single deployment,
  because once the bootstrapped tokens have been spent, they can't
  be spent again. But now we often do CD where we deploy stable,
  then deploy next gen, then increase block version, and do testing
  in-between each step. (For instance, in one case we had a bug where
  the next enclave version at block version 0 got stuck if there was
  more than one transaction in the block, so it would pass many forms
  of testing but not a slam test.)

The `mobilecoind-dev-faucet` was created to address the first problem,
because it can distribute any kind of token that mobilecoind can work
with, and it is not tied to the bootstrap process in any way.

This commit hopes to address the second problem. To do this, we want
to re-use some of the `mobilecoind-dev-faucet` internals.

* `mobilecoind-dev-faucet` already has a background worker thread that
  produces a continuous stream of "small" TxOut's that can be spent
  in parallel.
* This is exactly what `slam` needs to work, and what it previously
  was getting from the bootstrapped ledger, which is why it was tied
  intimately to the bootstrap. If we make it get TxOut's from the
  `mobilecoind-dev-faucet` utxo queue instead, then we can make it
  work as many times as we like in a deployed network.

Unfortunately, the `mobilecoind-dev-faucet` cannot conduct the slam
by submitting requests to `mobilecoind` -- `mobilecoind` is not
optimized for throughput, and it only handles one request at a time.
`fog-distro` is the only tool that spawns multiple threads to be
able to submit as many Tx's as possible in parallel.

So, this moves a bunch of that logic into the nascent `slam` module
in `mobilecoind-dev-faucet`. The benefit though is that, it will
allow us to delete the other half of `fog-distro`, the part where it
reads UTXOs from the bootstrapped ledger, which was the most
problematic.

We cannot delete `fog-distro` yet without updating CI and CD, but
if we can get a passing CD using this tool instead of `fog-distro`,
then we could delete `fog-distro` and likely end up with less
lines of code overall. If in the future `mobilecoind` is changed to
allow for increased parallelism then we could look at reducing
the code footprint in the `mobilecoind-dev-faucet` at that time.

This tool may allow us to do an improved release testing process
for 1.3:

* We will be able to do load testing at every block version
* Clients can request any account to be funded for testing, as long
  as they have the  faucet URL, rather than every client needing to
  be handed pre-funded accounts.